mke2fs.conf: enable the metadata_csum_seed and orphan_file features by default
The orphan_file feature is a compat feature which is first supported
in the Linux kernel in 5.15.

The metadata_csum_seed feature is an incompat feature which is first
supported in the Linux kernel starting in the 4.4 kernel and e2fsprogs
1.43.

Note: In order for grub to be able to accept a file system with
metadata_csum_seed, it must have commit 7fd5feff97c4 ("fs/ext2: Ignore
checksum seed incompat feature").  This commit landed in July 2021,
which just missed the Grub 2.06 release in June 2021.  Grub 2.11 was
slated for "H1'2022" but unfortunately it has not yet been released as
of this writing.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>