Disable metadata_csum_seed and orphan_file by default for Debian
There are a lot of debian packages that assume that mkfs.ext4 for
Debian N will work on Debian N-1 (and sometimes N-2, N-3, etc.).  This
is a bad assumption, but it's too late to fix this for Debian
Bookworm.

So revert making these feature a default for now.  We will enable it
again post Bookworm.

Signed-off-by: Theodore Ts'o <tytso@mit.edu>