debugfs: fix endless loop when calling rdump on fs where extended attribute value conflicts with key.
When recursively dumping the content of an ext2 filesystem where
extended attribute value conflicts with key using "rdump / /tmp/out ",
debugfs get's in an endless loop.

This is due to the fact that we only break out of the while(1) loop if
the 'got' value is 0, which never happens when an error is encountered.

The applied fix mimics what happens when retval is non-zero and returns.

This can be tested with an existing test sample
(tests/f_ea_value_crash/image.gz).