blkid: Add support for > 64k cluster NTFS volumes.
Windows 10 Creators Edition introduced larger cluster sizes for NTFS,
but since the number of sectors per cluster is stored as an 8-bit
integer and 128 is the largest power of 2 that can be represented as an
unsigned 8-bit integer, larger cluster sizes are encoded as a bitshift
value / power of two counted down from the value 255.
I.e. past the sectors_per_cluster value 128 the actual number of sectors
per cluster is calculated as 2^(256 - sectors_per_cluster).